// types/proxy.d.ts
export interface ProxyDetails {
  [key: string]: any;
  proxy: string;
  latency: string;
  last_check: string;
  type: string;
  region: string;
  city: string;
  country: string;
  timezone: string;
  latitude: string;
  longitude: string;
  anonymity: string;
  https: string;
  status: string;
  private: string;
  lang: string;
  useragent: string;
  webgl_vendor: string;
  webgl_renderer: string;
  browser_vendor: string;
  username: string;
  password: string;
}